made pre-merge code corrections
