+++ src/components/Home/index.tsx
@@ -1,20 +1,19 @@
-import ListeningTo from "../ListeningTo";
+import StreamRadio from "../StreamRadio";
 import CardBanners from "../CardBanners";
 import Channels from "../Channels";
 import NavigationButtons from "../NavigationButtons";
 import "./styles.scss"
 
-const Home = () => {
-  return (
-    <>
-      <CardBanners />
-      <NavigationButtons />
-      <div className="home-lower-container">
-        <Channels />
-        <ListeningTo />
-      </div>
-    </>
-  );
-};
+const Home = () => (
+  <>
+    <CardBanners />
+    <NavigationButtons />
+    <div className="home-lower-container">
+      <Channels />
+      <StreamRadio />
+    </div>
+  </>
+)
+
 
 export default Home;
