+++ src/components/StreamerCard/styles.scss
@@ -1,0 +1,80 @@
+.ch-streamers {
+    width: 100%;
+    height: 467px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+
+    .ch-streamer-card {
+        display: flex;
+        justify-content: space-between;
+        flex-direction: row;
+        width: 100%;
+        height: 52px;
+
+        .ch-card-left {
+            display: flex;
+            flex-direction: row;
+            gap: 20px;
+
+            .ch-streamer-icon {
+                height: 100%;
+                width: 52px;
+
+                img {
+                    width: 100%;
+                }
+            }
+
+            .ch-streamer-info {
+                display: flex;
+                flex-direction: column;
+                justify-content: space-around;
+
+                .ch-streamer-title {
+                    height: 21px;
+                    display: flex;
+                    gap: 10px;
+
+                    .ch-streamer-name {
+                        font-size: 16px;
+                        font-weight: 600;
+                    }
+
+                    .ch-streamer-popular-tag {
+                        display: flex;
+                        justify-content: center;
+                        align-items: center;
+                        width: 82px;
+                        background-color: #35BC29;
+                        border-radius: 9px;
+
+                        span {
+                            font-size: 8px;
+                            font-weight: 800;
+                            text-transform: uppercase;
+                        }
+                    }
+                }
+
+                .ch-streamer-stream-tag {
+                    font-size: 14px;
+                    font-weight: 400;
+                }
+            }
+        }
+
+        .ch-card-right {
+            display: flex;
+            align-items: center;
+            gap: 10px;
+
+            .ch-rec-icon {
+                height: 10px;
+                width: 10px;
+                background: #FB3000;
+                border-radius: 50%;
+            }
+        }
+    }
+}
