--- src/components/Streams/index.tsx
@@ -1,7 +1,0 @@
-import "./styles.scss";
-
-const Streams = () => {
-  return <div className="streams">Streams</div>;
-};
-
-export default Streams;
