--- src/components/Streams/styles.scss
@@ -1,5 +1,0 @@
-@import "./../../styles/variables";
-
-.streams {
-  color: $font-primary-color;
-}
