+++ src/data/images.ts
@@ -8,7 +8,7 @@ import banner7 from './../assets/card-banner7.png';
 import banner8 from './../assets/card-banner8.png';
 import banner9 from './../assets/card-banner9.png';
 
-const streamBanners: string[] = [
+const CardBannersList: string[] = [
     banner1,
     banner2,
     banner3,
@@ -20,5 +20,5 @@ const streamBanners: string[] = [
     banner9,
   ];
   
-  export default streamBanners;
+  export default CardBannersList;
